added enums so we can select magical girl types in the unity inspector
