+++ Assets/Scripts/MagicalGirls/MagicalGirlController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e54f4035741ba54f8b184301609448a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
