+++ Assets/Scripts/MagicalGirls/MagicalGirlPlaceholderController.cs
@@ -4,10 +4,14 @@ using UnityEngine;
 
 public class MagicalGirlPlaceholderController : AbstractMagicalGirlController
 {
+	[SerializeField] protected MagicalGirlTypeEnums.HappyMagicalGirlTypes happyType;
+	[SerializeField] protected MagicalGirlTypeEnums.AngryMagicalGirlTypes angryType;
+
 	public MagicalGirlPlaceholderController () : base()
 	{
-		angryState = new DevAngryMagicalGirl(this);
-		happyState = new BasicHappyMagicalGirl(this);
+		happyState = MagicalGirlTypeEnums.ConvertHappyType(this.happyType, this);
+		angryState = MagicalGirlTypeEnums.ConvertAngryType(this.angryType, this);
+
 		if (isAngryAtStart)
 			magicalGirlState = angryState;
 		else magicalGirlState = happyState;
