+++ Assets/Scripts/MagicalGirls/MagicalGirlTypeEnums.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb026b9ee64181244be83a9d1b1e93a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
